backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
celery==5.4.0
redis==5.1.1
boto3==1.35.32
anthropic==0.34.2
google-generativeai==0.8.3
pillow==11.2.1
pillow-heif==0.22.0
pymupdf==1.24.10
authlib==1.3.2
